Move to cling 0.0.1-alpha.1
